Move SSE client to client.py
So that SSE client can use AbstractAuth (which fixes a mypy issue)
